evaluation: add StartTime to after callbacks
